requirements.txt
Django<5
psycopg2
gunicorn
django-cors-headers
whitenoise
dj-database-url
django-bmemcached
fonttools
google-cloud-translate==2.0.1